+++ ospray/lights/registration.cpp
@@ -1,7 +1,8 @@
-// Copyright 2020 Intel Corporation
+// Copyright 2020-2021 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
 #include "AmbientLight.h"
+#include "CylinderLight.h"
 #include "DirectionalLight.h"
 #include "HDRILight.h"
 #include "PointLight.h"
@@ -14,6 +15,7 @@ namespace ospray {
 void registerAllLights()
 {
   Light::registerType<AmbientLight>("ambient");
+  Light::registerType<CylinderLight>("cylinder");
   Light::registerType<DirectionalLight>("distant");
   Light::registerType<HDRILight>("hdri");
   Light::registerType<PointLight>("sphere");